Moved CustomCanvasLabel.py to its own py_package
